edit dashboard navbars
