+++ src/data/navbarStudent.js
@@ -1,0 +1,15 @@
+import Home from '../assets/icons/home.png';
+import Lesson from '../assets/icons/lessons.png';
+import Messages from '../assets/icons/messages.png';
+import Notification from '../assets/icons/notification.png';
+import Payment from '../assets/icons/payment.png';
+
+const navbarItemsStudent = [
+    { src: Home, text: 'Home', link: '' },
+    { src: Lesson, text: 'My lessons', link: 'lessons' },
+    { src: Notification, text: 'Notifications', link: 'notifications' },
+    { src: Messages, text: 'Messages', link: 'messages' },
+    { src: Payment, text: 'Payments', link: 'payments' },
+];
+
+export default navbarItemsStudent;
