+++ src/data/navbarTeacher.js
@@ -1,0 +1,17 @@
+import Classes from '../assets/icons/classes.png';
+import Home from '../assets/icons/home.png';
+import Messages from '../assets/icons/messages.png';
+import Notification from '../assets/icons/notification.png';
+import Schedule from '../assets/icons/schedule.png';
+import Student from '../assets/icons/student.png';
+
+const navbarItemsTeacher = [
+    { src: Home, text: 'Home', link: '' },
+    { src: Classes, text: 'My classes', link: 'classes' },
+    { src: Student, text: 'My students', link: 'students' },
+    { src: Schedule, text: 'Schedule', link: 'schedule' },
+    { src: Notification, text: 'Notifications', link: 'notifications' },
+    { src: Messages, text: 'Messages', link: 'messages' },
+];
+
+export default navbarItemsTeacher;
